% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoloc.R
\name{button_geoloc}
\alias{button_geoloc}
\title{Geolocation button}
\usage{
button_geoloc(inputId, label, icon = NULL, width = NULL, ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{icon}{An optional \code{\link{icon}} to appear on the button.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link{validateCssUnit}}.}

\item{...}{Named attributes to be applied to the button or link.}
}
\value{
The user geolocation
}
\description{
Geolocation button
}
\examples{
library(shiny)
ui <- fluidPage(
  button_geoloc("foo", "bar")
)


}
