% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoconnectCSV.R
\name{geo_connect_csv}
\alias{geo_connect_csv}
\title{Google Geocoding API connector for CSV files}
\usage{
geo_connect_csv(csv_filepath)
}
\arguments{
\item{csv_filepath}{A csv file path}
}
\value{
A data.frame containing the resulting \code{longitude/s}, \code{latitude/s} and \code{address/es}
}
\description{
Takes a csv file path (with the required columns \code{Address} and/or \code{Latitude} and
 \code{Longitude}) as a parameter and creates a vector with the information that will be passed to the 
\code{geo_connect_vector} function that will connect with the Google Geocoding API and get the corresponding output.
}
\examples{
#geo_connect_csv(your_csv_filepath)


}
\references{
Google Geocoding API - \url{https://developers.google.com/maps/documentation/geocoding/intro}
}

