% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoconnectVector.R
\name{geo_connect_vector}
\alias{geo_connect_vector}
\title{Google Geocoding API connector for character vectors}
\usage{
geo_connect_vector(address)
}
\arguments{
\item{address}{A character vector with of address/es and/or some coordinates (latitude and longitude)}
}
\value{
A data.frame containing the resulting \code{longitude/s}, \code{latitude/s} and \code{address/es}
}
\description{
Loops over the vector of addresses and/or coordinates, passes each one of them as a parameter 
to an internal function that connects to the \code{Google Geocoding API} via URL and gets the output. Then it 
returns the output of the request into a formatted dataframe with the geocoding information.
}
\examples{
geo_connect_vector(c("Madrid", "Skattegarden 29B, Linkoping"))
geo_connect_vector(c("40.41678, -3.70379", "28.49634, -16.32415"))
geo_connect_vector(c("40.41678, -3.70379", "Skattegarden 29B, Linkoping"))

}
\references{
Google Geocoding API - \url{https://developers.google.com/maps/documentation/geocoding/intro}
}

