\name{geoloc-package}
\alias{geoloc-package}
\alias{geoloc}
\docType{package}
\title{
\packageTitle{geoloc}
}
\description{
\packageDescription{geoloc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{geoloc}
\packageIndices{geoloc}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{geoloc}

Maintainer: \packageMaintainer{geoloc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
