% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoloc.R
\docType{package}
\name{geoloc}
\alias{geoloc}
\alias{geoloc-package}
\title{A package for connecting to Google Geocoding API}
\description{
The \code{geoloc} package provides an interface to connect with the \code{Google Geocoding API} with 
functions that convert adresses into coordinates and vice versa.
}
\details{
The package provides the following functions:
\itemize{
   \item \code{Func1}: 
   \item \code{Func2}: 
}
}
\author{
\itemize{
     \item Teno González Dos Santos - \url{https://github.com/tenoglez}
     \item Josué Álvarez Robles - \url{https://github.com/ejarkm}
     \item Jose Luis López Ruiz - \url{https://github.com/pepelu8000}
   }
}
\references{
Google Geocoding API - \url{https://developers.google.com/maps/documentation/geocoding/intro}
}

