% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoloc.R
\name{onload_geoloc}
\alias{onload_geoloc}
\title{Geolocation at launch}
\usage{
onload_geoloc()
}
\value{
The user geolocation
}
\description{
Geolocation at launch
}
\examples{

library(shiny)
ui <- fluidPage(
  onload_geoloc()
)

}
