% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtfismyip.R
\name{wtfismyip}
\alias{wtfismyip}
\title{WTF is my IP}
\usage{
wtfismyip(safe = TRUE)
}
\arguments{
\item{safe}{wether or not to include the F words in result names}
}
\value{
a list with IP Address, Location, Hostname, ISP,
    TorExit, and Country code.
}
\description{
WTF is my IP
}
